services/user-managment/requirements.txt
Django==5.2.8
djangorestframework==3.15.0
psycopg[binary]==3.2.13
